Fixed test for diag in Torch frontend

+++ ivy_tests/test_ivy/test_frontends/test_torch/test_miscellaneous_ops.py
@@ -338,7 +338,7 @@ def test_torch_diag(
 ):
     dtype, values = dtype_and_values
 
-    values = np.asarray(values)
+    values = np.asarray(values, dtype=dtype)
 
     helpers.test_frontend_function(
         input_dtypes=dtype,
